Updated messenger to work with slightly different storage server API and changed swarm requests to go through storage server
