+++ config/default.json
@@ -3,8 +3,7 @@
   "localUrl": "localhost.loki",
   "cdnUrl": "random.snode",
   "localServerPort": "8081",
-  "messageServerPort": "8080",
-  "swarmServerPort": "8079",
+  "snodeServerPort": "8080",
   "disableAutoUpdate": false,
   "openDevTools": false,
   "buildExpiration": 0,
