+++ main.js
@@ -147,8 +147,7 @@ function prepareURL(pathSegments, moreKeys) {
       serverUrl: config.get('serverUrl'),
       localUrl: config.get('localUrl'),
       cdnUrl: config.get('cdnUrl'),
-      messageServerPort: config.get('messageServerPort'),
-      swarmServerPort: config.get('swarmServerPort'),
+      snodeServerPort: config.get('snodeServerPort'),
       localServerPort: config.get('localServerPort'),
       certificateAuthority: config.get('certificateAuthority'),
       environment: config.environment,
